Add "startup" options to start Rserve processes.  It should work remotely (through ssh), however, because of firewalls, etc shutting down needs to be done on localhost.
Had to add the "full" path to the startup command in the xml cofiguration.  The sample has been updated, but it's basically what is echoed when the Rserve process is started manually through the R command interface.  I'm not happy about this, but I don't see a good siimple way around it given the firewalls and different architectures.

git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/icb-commons/RUtils@15111 60cd2ccc-2b05-0410-9566-da273a940bff